feat(helm): add --force flag to 'helm install'
When 'helm install --force' is run it will use be the net equivalent of
using kubectl apply --overwrite to apply the chart which will overwrite
any existing objects in the cluster. This is an unsafe operation but is
useful for people that want to update a deployment that wasn't
previously owned by helm without completely deleting and creating it
from scratch.